Rename backup infrastructure module
